fixup typecast

+++ src/tool_writeout.c
@@ -573,7 +573,7 @@ static const char *outtime(const char *ptr, /* %time{ ... */
         curl_off_t val;
         curlx_str_number(&timestr, &val, TIME_T_MAX);
         cnow.tv_sec = (time_t)val;
-        cnow.tv_usec = val % 1000000;
+        cnow.tv_usec = (unsigned int)(val % 1000000);
       }
     }
 #endif
